+++ src/main/java/com/example/fashioncoordinator/admin/domain/BrandPriceInfo.java
@@ -1,4 +1,4 @@
-package com.example.fashioncoordinator.api.admin;
+package com.example.fashioncoordinator.admin.domain;
 
 import lombok.Builder;
 import lombok.Getter;
