+++ src/main/java/com/example/fashioncoordinator/core/api/response/LowestPriceBrandProductResponseDto.java
@@ -1,8 +1,8 @@
-package com.example.fashioncoordinator.api.customer.response;
+package com.example.fashioncoordinator.core.api.response;
 
 import com.example.fashioncoordinator.db.ProductEntity;
 import com.example.fashioncoordinator.enums.ProductCategory;
-import com.example.fashioncoordinator.serializer.NumberWithCommaSerializer;
+import com.example.fashioncoordinator.core.api.serializer.NumberWithCommaSerializer;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import io.swagger.v3.oas.annotations.media.Schema;
