+++ src/main/java/com/example/fashioncoordinator/core/api/response/LowestPriceCombinationResponseDto.java
@@ -1,22 +1,27 @@
-package com.example.fashioncoordinator.api.customer.response;
+package com.example.fashioncoordinator.core.api.response;
 
 import com.example.fashioncoordinator.enums.ProductCategory;
-import com.example.fashioncoordinator.serializer.NumberWithCommaSerializer;
+import com.example.fashioncoordinator.core.api.serializer.NumberWithCommaSerializer;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 
 @Getter
 @Builder
-public class HighestLowestPriceBrandResponseDto {
+@EqualsAndHashCode
+public class LowestPriceCombinationResponseDto {
 
     @Getter
     @Builder
-    @Schema(name = "상품:브랜드,가격")
+    @EqualsAndHashCode
     public static class ProductResponseDto {
+
+        @JsonProperty("카테고리")
+        private ProductCategory category;
 
         @JsonProperty("브랜드")
         private String brand;
@@ -27,13 +32,12 @@ public class HighestLowestPriceBrandResponseDto {
         private int price;
     }
 
-    @JsonProperty("카테고리")
-    private ProductCategory category;
+    @JsonProperty("상품")
+    private List<ProductResponseDto> productList;
 
-    @JsonProperty("최저가")
-    private List<ProductResponseDto> lowestPriceProductList;
-
-    @JsonProperty("최고가")
-    private List<ProductResponseDto> highestPriceProductList;
+    @JsonProperty("총액")
+    @Schema(implementation = String.class)
+    @JsonSerialize(using = NumberWithCommaSerializer.class)
+    private int totalPrice;
 
 }
