+++ src/main/java/com/example/fashioncoordinator/core/domain/ProductService.java
@@ -1,9 +1,9 @@
-package com.example.fashioncoordinator.api.customer;
+package com.example.fashioncoordinator.core.domain;
 
-import com.example.fashioncoordinator.api.customer.response.HighestLowestPriceBrandResponseDto;
-import com.example.fashioncoordinator.api.customer.response.HighestLowestPriceBrandResponseDto.ProductResponseDto;
-import com.example.fashioncoordinator.api.customer.response.LowestPriceBrandProductResponseDto;
-import com.example.fashioncoordinator.api.customer.response.LowestPriceCombinationResponseDto;
+import com.example.fashioncoordinator.core.api.response.HighestLowestPriceBrandResponseDto;
+import com.example.fashioncoordinator.core.api.response.HighestLowestPriceBrandResponseDto.ProductResponseDto;
+import com.example.fashioncoordinator.core.api.response.LowestPriceBrandProductResponseDto;
+import com.example.fashioncoordinator.core.api.response.LowestPriceCombinationResponseDto;
 import com.example.fashioncoordinator.db.ProductEntity;
 import com.example.fashioncoordinator.db.ProductCustomRepository;
 import com.example.fashioncoordinator.db.ProductJpaRepository;
